requirements.txt
streamlit==1.38.0
numpy==1.26.4
pypdf==4.0.2
groq==0.6.0
python-dotenv==1.0.1
pillow==10.3.0
watchdog==4.0.0

